Initial support for CDL generation.
This is the first, basic, support for generating CDL.  It
currently only works on Fedora guests.  Basically the idea
is that, given a disk image and a piece of IDL, you can:

1)  Edit the diskimage (using libguestfs) to allow "passwordless" access
2)  Start up the guest
3)  Run the appropriate command to get the list of packages
4)  Shutdown the guest, restore image to previous state.

It is currently a bit ugly in that there are lots of FIXME
comments all over the place, and I don't love the fact that
you need to pass in both an IDL and a diskimage, but I'm
not quite sure how much better we can do.

Signed-off-by: Chris Lalancette <clalance@redhat.com>