Support for Windows 2008.
This is the initial support for Windows 2008.  This required
quite a few fixes:

1)  I wasn't properly extracting the El-Torito boot image, due
to some alignment issues.  This is now fixed.
2)  Windows 2008 uses a UDF filesystem instead of ISO.  Because
of that, we can no longer use fuseiso to mount the image.  Instead,
use libguestfs to extract the data, which seems to work for both
ISO or UDF.
3)  Windows 2008 is very picky about command-line arguments to
mkisofs.  Very carefully craft them here.

With this in place, oz is able to take a Windows 2008 ISO and
generate a new Windows 2008 ISO, with no modifications.  The
modifications/unattended.xml file still need to be added.

Signed-off-by: Chris Lalancette <clalance@redhat.com>