Redo the logic for when an install is complete.
In particular, if we see that the info has gone to
VIR_DOMAIN_SHUTOFF, that is a pretty good indication that the
install is complete, so we bail out and assume completion.
Otherwise, if the state is VIR_DOMAIN_RUNNING or
VIR_DOMAIN_BLOCKED, we assume that the install is still
running and continue through the loop.  Otherwise, we assume
failure and bail out.

Signed-off-by: Chris Lalancette <clalance@redhat.com>