Add the cdl-nc binary to the git repository.
The reason we do this is for the maximum possible portability.
That is, if you happen to build cdl-nc on a newer platform,
it very well may not work on an older platform.  Instead, I
have hand-compiled this version on Red Hat Linux 7.3 i386
(from 2002) like so:

gcc -g -Wall -static cdl-nc.c -o cdl-nc

This should really work on all modern Linux, and any Linux
back to Red Hat Linux 7.3.

Signed-off-by: Chris Lalancette <clalance@redhat.com>