Remove the check_url function.
It was a completely unnecessary step.  Not only was it
manually handling redirects (which is much better handled
by urllib2), it didn't support FTP and it also fractured
our URL handling into two places.  With the recent change
to urllib2, we don't need it at all anymore, so just
remove it here.

Signed-off-by: Chris Lalancette <clalance@redhat.com>