Massive refactoring.
This is a large refactoring of the code.  The impetus is mostly
for code sharing between the various RedHat-based OSs; they mostly
need to do the same things, so I wanted to share between them.
To that end, I introduce a RedHat.RedHatCDGuest, that all of
[Fedora, FedoraCore, RHEL_6, RHEL_5, RHEL_4, RHEL_3] are subclasses
of.

Additionally, this patch changes from a stateful libvirt domain
(conn.defineXML/dom.create) to using transient domains.  The
advantage is that we'll never have stale libvirt domains hanging
around; they will always be gone after installation.  At the
end of installation, we actually dump the libvirt XML out to a
file, so that it can be inspected (or used by other tools).

Signed-off-by: Chris Lalancette <clalance@redhat.com>