Make a rough guess if we will run out of diskspace.
Right now, this only detects if we will run out of diskspace while
downloading the install media.  We may also want to add a check
to make sure there will be enough space to extract the install media
and create a new one, and also enough space to actually do the
install.

Signed-off-by: Chris Lalancette <clalance@redhat.com>