Use a dummy knownhosts file for ssh/scp.
This is so that we avoid warnings to stdout like:

05:54:82:da:91:20:29:64:d4:aa:28:ba:17:ee:06:9e
@    WARNING: REMOTE HOST IDENTIFICATION HAS CHANGED!
@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
Add correct host key in /root/.ssh/known_hosts to get rid of this message.
IT IS POSSIBLE THAT SOMEONE IS DOING SOMETHING NASTY!
It is also possible that the RSA host key has just been changed.
Keyboard-interactive authentication is disabled to avoid man-in-the-middle attacks.
Offending key in /root/.ssh/known_hosts:137
Password authentication is disabled to avoid man-in-the-middle attacks.
Please contact your system administrator.
Someone could be eavesdropping on you right now (man-in-the-middle attack)!
The fingerprint for the RSA key sent by the remote host is

Signed-off-by: Chris Lalancette <clalance@redhat.com>