Fix up error detection for fetching the boot media.
As pointed out by Marek Goldmann while at FUDCon, we
need to have better error messages if we can't reach the
destination.  Here, we clean up the error handling in
get_original_media() so that if urllib2 throws an
error, we let it bubble up the stack.  We also now
check to see if Content-Length exists before trying
to access the key, and throw an exception if not.

Signed-off-by: Chris Lalancette <clalance@redhat.com>