Use a pipe to extract ISO contents.
I've now figured out the problem with hanging pipes;
essentially if your output gets > 64k, the subprocess.PIPE
will fail.  The workaround is to open your own file descriptors
and bypass the python subprocess.PIPE.  Now that we have this
fixed, we can get rid of the temporary file we were using
to hold the ISO tar data.  This should both save space and
significantly speed up the extraction.

Signed-off-by: Chris Lalancette <clalance@redhat.com>