Make sure to kill off the tar subprocess if tar_out fails.
What was happening was that if guestfs.tar_out failed, the
entire thing would hang up.  If we do see an exception
for guestfs.tar_out, make sure to kill off the tar subprocess
and then re-raise the error.  Along with this, refactor the
rest of the copy_iso code to clean up after itself.

Signed-off-by: Chris Lalancette <clalance@redhat.com>