Consolidate the Ubuntu logic.
I realized that all of the Ubuntu logic was essentially doing
the same thing.  I further realized that instead of going in
and trying to "fix-up" the isolinux.cfg files on the CD for
the automated boot, it would be much easier to just replace it
entirely.  That's what this patch does.  Because of this,
I was able to remove lots of redundant code in the Ubuntu
codepaths.

Signed-off-by: Chris Lalancette <clalance@redhat.com>