Improve subprocess usage.
In particular, make sure that we use temporary fds for
stdout and stderr to avoid the nasty subprocess hang problem.
Along with this, make sure all uses of subprocess in oz
(with the exception of one very special case) use the wrapper
so that they don't run into the problem.

Signed-off-by: Chris Lalancette <clalance@redhat.com>