Disallow oddly configured servers for RedHat-based yum OSs.
It is possible to configure some http servers to not allow
fetching of byte-ranges (which is allowed by the HTTP/1.1
spec).  However, anaconda in Fedora and RHEL don't honor
this, which essentially causes the build to fail.  We can't
really fix this in released anaconda, so the best that we
can do is to detect the situation prior to launching the
install and raising an exception.

Signed-off-by: Chris Lalancette <clalance@redhat.com>