Another fix for servers that don't support byte ranges.
The comment in the code explains this pretty well, but
to avoid fetching different packages from different mirrors,
have Oz pick a single mirror (in the case of a redirect) up-front
and verify that it supports byte-ranges.  If it doesn't, try to
pick another mirror (up to 5 times) to try to find one that
does.  If no mirrors that support byte ranges can be found,
give up.

Signed-off-by: Chris Lalancette <clalance@redhat.com>