Add the ability to override the libvirt bridge.
By specifying a key in the oz configuration file, you can
now override the bridge that Oz will use for installing
guests.  By default Oz will continue to try to detect
and use the first NAT bridge it finds on the system; this
just gives users a little more flexibility.

Signed-off-by: Chris Lalancette <clalance@redhat.com>