Revamp ssh key generation.
There were a couple of problems with the current ssh key
generation:

1)  If run in a SELinux confined environment, the subprocess
call to ssh-keygen would be denied
2)  We used a subprocess call to generate the key
3)  If lots of builds were run, they would use up a lot of
entropy from the machine
4)  We aren't really increasing security by generating a key
every time (pointed out by Ian Mcleod)

To alleviate these concerns, we generate a key if, and only if,
it doesn't exist in a pre-defined location.  If it does exist,
we just re-use that key.  Additionally, when installing through
RPM, we have the RPM spec file generate the key.  That means
that in the normal case, we generate a key once when oz is
installed and never again.  If a user happens to download oz
from source and run it, we'll generate a key using python-crypto,
but only during the first install.  Subsequent installs will
use this generated key.  These points address problems 2, 3,
and 4.  We still potentially have a problem with 1, but the user
would need to a) download from source and b) use oz from a
SELinux confined environment.  This situation is oddball enough
that I think we can ignore it for now.

Signed-off-by: Chris Lalancette <clalance@redhat.com>