Remove the use of paramiko/python-crypto to generate keys.
There were multiple problems with using python-crypto to
generate the openssh keys:

1)  You needed a very new version of python-crypto that
doesn't exist in RHEL-6
2)  There was hand-crafted code to convert the PEM format
to something that ssh understands
3)  The python-crypto code would crash under certain
circumstances
4)  python-crypto has hand-written implementations of
various encryption algorithms

Instead, go back to using a subshell to generate keys
using ssh-keygen.  Because we are generating the key
during RPM install, we'll still have most of the benefit
as before, but without the problems that python-crypto
brings.

Signed-off-by: Chris Lalancette <clalance@redhat.com>