Add a shorter alive interval.
Certain firewall configurations might reap connections
that are inactive for a long period of time.  If a
long running command is fired off via ssh, and it produces
no output, the connection might get reaped as inactive.
To prevent this, send a keep-alive every 30 seconds to
keep the connection alive.

Signed-off-by: Chris Lalancette <clalance@redhat.com>