Start with a blank diskimage.
The previous generate_diskimage code laid down a fake
partition table on the disk, but it did it in a completely
bogus way.  Also, for most operating systems, starting with
a blank disk is just fine.  The only one that I know of
that currently has a problem is Windows 2000.  For that,
use pyparted to create a bogus partition table that will
get blown away.

Signed-off-by: Chris Lalancette <clalance@redhat.com>