Bugfix in wait_for_guest_boot.
wait_for_guest_boot did something really stupid;
it would continue to wait around for a domain to
check in, even if the domain in question had gone
away.  It was essentially waiting for no reason.

To fix this, we pass in the domain pointer and then
check the domain info every time we leave accept
and don't have data on the read socket.  As long as
the domain is still alive, domain.info() won't throw
an exception and we continue to wait.  If the domain
has crashed or otherwise gone away, domain.info()
will throw an exception and we will stop waiting
unnecessarily.

Along with this, we also change the select timeout to
1 second so that we wake up more often to check on the
domain status.  We still only print the debug information
every 10 seconds.

Signed-off-by: Chris Lalancette <clalance@redhat.com>