Push the setting of self.url down into RedHatCDGuest.
pylint was complaining that self.url was never set anywhere
in RedHatCDGuest.  What it couldn't see was that the classes
that subclassed RedHatCDGuest were always setting self.url.
However, that is a rather tenuous situation; if you happen
to subclass RedHatCDGuest and forget to set self.url, then
pylint will be correct.  To fix up this apparent discrepancy,
push the setting of self.url down into the __init__ for
RedCDGuest.

Signed-off-by: Chris Lalancette <clalance@redhat.com>