Install on Fedora 15 systems and allow customization
Fedora 15 systems use a version of cronie that is tailored towards systemd.
As a result, the default RedHat.py cron handling doesn't work on Fedora 15.

This patch has been tested with cron disabled to start (where it was returned
to disabled) and cron enabled to start (where it remained enabled) on both
F15 and RHEL6.1.

Signed-off-by: Steven Dake <sdake@redhat.com>