Implement secret passing during wait_for_guest_boot.
This is a feature to prevent mistaken connections to
the callback port tricking oz into thinking that the
guest is up.  We essentially use the UUID of the guest
as a shared secret, and when connecting back to oz
the guest passes the secret.  If it matches, oz knows
the guest is up.  If it doesn't, oz continues waiting.

Signed-off-by: Chris Lalancette <clalance@redhat.com>