Add a flag to do forced root password checking.
By default, we do nothing to enforce a root password,
and use the ozrootpw as necessary.  However, this may
not be desirable in some situations, so we add a flag
to TDL parsing to enforce a root password.  If this
flag is set, then we raise an error if the root password
is missing.

Signed-off-by: Chris Lalancette <clalance@redhat.com>