Prototype implementation of the "proxy" style download.
This is where we proxy the yum requests from the guest down
to the host machine to be resolved.  The benefit here is
that the packages only need to be available to the controlling
machine; they'll be uploaded via the ssh tunnel to the
guest as needed.

The problem is that this current implementation is kind of
complicated.  In particular, we need to start off a threaded
server to handle the proxy requests, and I'm not sure how
well that will interact with the already threaded imagefactory.
This requires more thought and experimentation.

Signed-off-by: Chris Lalancette <clalance@redhat.com>