Use libvirt capabilities to discover the virt type to use.
Instead of indirectly going through virt-what to discover
the virtualization type, instead just ask libvirt what it
supports.  Given that information, we can then make a decision
as to which virtualization type we prefer.  This gets rid
of a subprocess, gets rid of the virt-what dependency, and
goes right the source of things.

Signed-off-by: Chris Lalancette <clalance@redhat.com>