Ignore destroy failures if the domain went away.
There is a race inherent in shutdown_guest, in that between
the time that wait_for_guest_shutdown failed, and the
time that we go to destroy the domain, the domain may
have already gone away.  Handle that race here by only
raising an error if the domain is still there if
libvirt_dom.destroy fails.

Signed-off-by: Chris Lalancette <clalance@redhat.com>