Add file locking to the downloads.
This allows us to launch two oz installs at the same
time without tromping over each other.  Essentially, we
use file locking on the install media that we are downloading.
When two oz processes come along at the same time, the
second (or third, etc) one blocks waiting for the first
one to finish downloading.  Once it does, the others are
unblocked and allowed to proceed (at which point they
probably don't have to do the download anymore).

Signed-off-by: Chris Lalancette <clalance@redhat.com>