Rework mkdir_p to be more resistant to races.
Basically, we have to account for the fact that the directory
might have been created between the check (os.access) and the
actual makedirs call.  What that actually leads to is the
realization is that we should just do the os.makedirs(), and
suppress the error if the directory exists and is a directory.
All other problems get thrown as exceptions.

Signed-off-by: Chris Lalancette <clalance@redhat.com>