Support direct initrd injection.
This is a method of injecting files into a RedHat-style
OS without first downloading the boot.iso (for URL
style installs only).  Essentially we download the kernel
(vmlinuz) and ramdisk (initrd.img) from the remote server.
We then create a CPIO archive and gzip it, and append it
to the initrd that we downloaded.  Finally we use libvirt
direct kernel and initrd booting to run the install.

Signed-off-by: Chris Lalancette <clalance@redhat.com>