Remove most uses of os.access.
Code like:

if os.access(file, os.F_OK):
	os.unlink(file)

is inherently racy since the file might have disappeared
between the access and the unlink.  For the above to work
without error you would have to wrap a try around it.  But
in that case, you might as well wrap a try around the whole
thing and dispense with os.access completely.  This patch
does exactly that.

Signed-off-by: Chris Lalancette <clalance@redhat.com>