Change the announcement code to use serial instead of network.
Using the network for guest announcement (to discover
the IP address for the guest) is problematic in that
it requires the host firewall to be open.  While this
isn't generally a problem when using virbr0, if the
user has specified an external bridge the port won't
always be open.  This lead to nasty things like manipulation
of the host iptables.

To alleviate this problem, we instead change over to a serial
based approach.  After the guest properly configures the
default network route, it writes that data to the serial
port.  Since Oz had previously setup the serial port to
redirect to a network server hosted by qemu, Oz can just
poll the qemu network server to get the IP address.  Once
it sees properly formatted data Oz can parse it and
discover the guest IP address.

Signed-off-by: Chris Lalancette <clalance@redhat.com>