Include githash when building pre-release packages
A pre-release is any release where %{release} starts with 0.  When
building RPMs of pre-releases, a datestamp and githash are included in
the RPM %{release} string.  This allows to easily determine when and
what is included in the built packages.

v2: clalance: Change some of the code to be more pythonic
v3: jlaska: Let setup.py handle @RELEASE@

Signed-off-by: Chris Lalancette <clalance@redhat.com>