Allow oz to be run as non-root.
This works, with a lot of caveats.  As far as I know, they
are:

1)  The user that wants to run Oz must have access to the
libvirt system session (typically qemu:///system).  This
can be through permissions, through PackageKit, or through
any other supported libvirt mechanism, but
"virsh -c qemu:///system list" *must* work.
2)  The entire chain of directories leading up to the
place where the user wants to store the diskimage must
have rx bits (chmod o+rx /path/to/location).  This is because
libvirt will usually run the guest as qemu:qemu, meaning
that user has to be able to open the diskimage for writing.

Signed-off-by: Chris Lalancette <clalance@redhat.com>