Work around virtualenv issues when $PWD includes spaces
When running 'make virtualenv unittests' from a $PWD that includes
spaces, tests will fail to execute because bash is unable to understand
a script intepreter path that contains spaces.  To work around this
issue, if the intepreter specified by any virtualenv-provided scripts
includes a space, it is replaced by calling '/usr/bin/env $foo'.

For example ...
    #!"/this is a/path/with/spaces/bin/py.test

Would be replaced with...
    #!/usr/bin/env py.test

For more information, refer to
https://bugs.launchpad.net/virtualenv/+bug/241581.

Signed-off-by: Chris Lalancette <clalance@redhat.com>