Fix up the TDL RNG.
The last patch had a couple of problems.  The first problem
was that the various elements to a repository are individually
optional, while the RNG caused them to be optional as a block.

The second problem is that the rules for clientcert and
clientkey weren't laid out right.  In this new layout, I tried
to follow the rules from imcleod:

"You cannot have a clientkey without a clientcert.  All other
combinations of those two are valid."

The last problem is that we were totally missing the sslverify
and cacert tags.  Again, imcleod:

"If sslverify is yes you must have a cacert.  If it is no the
cacert is not required and, if present, is ignored."

RNG isn't quite expressive enough to get these 100% right, but
this patch at least fixes the unittests and gets most of the
above semantics in place.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>