Add a unittest for <persisted> repos
This patch also includes the exception text when failing to import the
oz module.  It's possible that the failure to import oz is the result of
missing libraries.  Displaying the exact ImportError exception will
highlight the cause of the failure.

Additionally, string_to_bool() error handling has been adjusted.
Previously, it would return None when supplied with bogus input.  This
patch now raises an exception, and corrects the corresponding unittest.