Silencing the curl command executed on the guest.
Added --silent to the guest curl command.  We already print a nice
debug statement if we need to set up a tunnel from the guest.  The
curl error code was getting logged and would be somewhat alarming to
the an administrator if they didn't have context into how we are determining
network connectivity.  The curl error is a non-exception case, and thus doesn't
need to be logged.