[PATCH] Implement SSL vhost support in ssh tunneling.
This patch creates a forwarding tunnel to the repository hosts.  A
temporary host entry is added as an alias in the guest /etc/hosts.
When the tunnel is finished being used, the alias is removed and any
persisted repositories are reverted to their original base urls.

Patch originally from Chris Alfonso.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>