Modify MANIFEST.in to fix RHEL6 RPM builds
RHEL6 uses a slightly older Python version (and associated libs) than recent Fedoras.
Without explicitly adding the oz/auto/* content to the MANIFEST the auto files don't
make it into the sdist and, as a result, are missing from the RPMS.

Adding this content to the manifest fixes the issue and doesn't break things on the
newer Python dists in Fedora.