guestfs_handle_setup: handle disappearing domains
The operation to look at running instances and check
if their names or disks conflict with what we are about
to do is inherently racy.  conn.listDomainsID() can return
a domain which disappears before the subsequent
lookupDomainByID can complete.  Instead of bombing out,
just log the fact that a domain disappeared on us and
continue on.

Original patch is from Ian Mcleod.

Signed-off-by: Chris Lalancette <clalance@redhat.com>