Add an fsync to the rmtree to reduce failures.
In particular, after doing an rmtree of a very large
ISO there can be a lot of metadata changes entering
writeback.  This can cause operations further along
(including libvirt ones) to fail.  To try and reduce
this pain, make sure to fsync on the parent of the
directory that we just deleted.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>