Properly specify kickstart file location for Fedora
According to the kickstart docs for every fedora release, the proper way
to specify a kickstart cdrom is with ks=cdrom:/ks.cfg.  See:
http://docs.fedoraproject.org/en-US/Fedora/13/html/Installation_Guide/s1-kickstart2-startinginstall.html
http://docs.fedoraproject.org/en-US/Fedora/14/html/Installation_Guide/s1-kickstart2-startinginstall.html
http://docs.fedoraproject.org/en-US/Fedora/15/html/Installation_Guide/s1-kickstart2-startinginstall.html
http://docs.fedoraproject.org/en-US/Fedora/16/html/Installation_Guide/s1-kickstart2-startinginstall.html
http://docs.fedoraproject.org/en-US/Fedora/17/html/Installation_Guide/s1-kickstart2-startinginstall.html

Fedora 17 has changed this behavior to require the syntax:

I have tested this patch on f16 and f17, but nothing earlier.

ks=cdrom:[device]:/ks.cfg

Signed-off-by: Steven Dake <sdake@redhat.com>