deb: Use dh2_python rather then cdbs to build Debian/Ubuntu packages
This patch updates the debian packaging to more appropriately match
what is expected in a multi-python environment.  While this only builds
python 2.x binaries, the previous debian oz packages didn't work at all on
modern Debian or Ubuntu systems because of the pyshared usage.  This resulted
in an error about TDL.py not being importable.

I also removed cdbs since it is not necessary to get the job done.

Signed-off-by: Steven Dake <sdake@redhat.com>