Save the startup links for both cron and ssh.
This is for error reporting.  The Ubuntu Desktop CDs
don't include the openssh-server package, so they
don't have /usr/sbin/sshd or /etc/init.d/ssh.  We already
had a check for the former, but in the course of handling
the exception for that we were generating a second
exception during cleanup because we were missing the
latter.  Save the startup links so that during teardown
we can use it without poking around in the disk image.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>