Fix up command ordering.
As pointed out by Ian Mcleod, we were storing the commands
to be run in a dictionary.  This meant that we weren't
honoring any ordering that the user intended.  Now,
XML doesn't really specify that the order that the
data is displayed is the order that the data is parsed
or stored.  However, in our particular context it is
pretty clear that the order the user specified is probably
the one they want the commands be run in.  Luckily, all
parsers that we know of parse in the textual order and
return data in that order.  In the future we might add
an "order" attribute to make this explicit, but this
patch fixes the immediate problem.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>