Switch GuestFactory to data-driven rather than code-driven.
This means that we have a dictionary that contains all
of the supported names and modules.  Then we use generic
code to generate the object by matching the dictionary
names with what the user requested.  While we are at it,
do the same thing for the list of distros; this means
that when we add support for a new distro, we don't
have to change GuestFactory anymore.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>