Loosen the check for RHEL-6 update numbers.
We have never had a case where a RHEL minor update has
broken the installer.  Thus, we can afford to accept any
version number within a major RHEL version.  This will make
it so that we don't have to release a new version of Oz
for every minor RHEL/CentOS update

Signed-off-by: Chris Lalancette <clalancette@gmail.com>