Quote apt-add-repository arguments to allow for non-ppa inputs
This allows for `deb http://example.com foo main` type of input. apt-add-repository requires it to be quoted. A PPA input can also be quoted without problems.
The commit allows people to add the `deb` bit to the TDL without having to add in the extra quotes around it in there.

Note: This would break things for people that have already used quotes in their TDL - That's why the strip() call is there

Signed-off-by: Chris Lalancette <clalancette@gmail.com>