Allow configuration to override the command used to generate ICICLE on Red Hat OSes
The default behavior with the default config file remains "rpm -qa"

This is to help with some work we are doing to integrate Oz/ImageFactory image
generation into koji.  The specific addition we made to the config file in this
case is:

[redhat]
icicle_command = "rpm -qa --qf '%%{NAME},%%{VERSION},%%{RELEASE},%%{ARCH},%%{EPOCH},%%{SIZE},%%{SIGMD5},%%{BUILDTIME}\n'"

Note that the extra %s are because SafeConfigParser treats % as a special character
which needs to be esacped by doubling it.