Make the lack of a libvirt bridge actionable.
As pointed out by Rich Jones, an exception that says:

ERROR: Could not find a viable libvirt NAT bridge, install cannot continue

is neither clear nor actionable.  Change this text to
explain better what the error is, and point to a wiki
page explaining more about what the problem is and
possible ways to remedy the problem.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>