Add in checks for Ubuntu desktop CDs.
Unfortunately, Ubuntu desktop CDs do not contain
the openssh-server.  Therefore, we cannot do
customization or ICICLE generation if the desktop
CD is in use (we can still do a basic install).
Instead of failing *after* the installation,
instead do a check right after unpacking the ISO
and throw an exception there.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>