Make file and command customization use less memory.
If we fetched files or commands via the URL mechanism
or through the base64 decoding mechanism, we used to
store all of the downloaded/decoded data in memory.
This is a huge memory hog, particularly for large
files.  Change this to use temporary files instead;
it will use much less memory, and is less code
to boot.

Note that commands still end up using the same amount
of memory, since they need to be read and executed
at some point.  But this is delayed until they are
actually used, and regardless this mechanism helps
for files.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>