Generalized version for adding extra folders to installation media
This patch allows the user to add extra folders to the
installation media. This is useful for windows use case, as it
allows to put the $OEM$ folder for customization. In other
environments, it can be used for placing extra content that needs
to be loaded from the installation media.