use unique name when creating temporary storage pool
If Oz is being used concurrently with two distinct directories and target
images it can still fail because of duplication of the temp storage pool
name (oztempdir).  There is code elsewhere in Guest.py that tries to detect
if a pool already exists for the same directory, independent of the name.

So, just create a unique name each time.