Make sure to sleep while doing ssh connection test.
If an ssh connection test failed very quickly, we could
burn through all of our tries before sshd in the guest
could ever had a chance to start.  Instead, make sure we
do a sleep in the loop if the ssh command very fast.
Also add some extra debug statements so it is easier to
debug this issue.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>