Speed up guest announcement on newer OSs.
A newer OS that uses NetworkManager can significantly
speed up guest announcement by using dispatch scripts.
These are scripts that execute when a NetworkManager
event happens, like eth0 getting connected.  Note that
we still leave the cron job in place; in the worst
case, the Oz process running on the host may miss
this announcement, and the cron job will make sure that
it gets seen (on average, 30 seconds later).

Signed-off-by: Chris Lalancette <clalancette@gmail.com>