new: Module to install FreeBSD 10.0.
It uses the same aproach that is used with Windows installation, where the iso is modified to incorporate a rules file. Bsdinstall reads such a file (from /etc/) and install the new system according to the rules and scripts found inside. After installing the base groups it's possible to install additional packages specified in the tdl. Running extra commands is not yet possible.

Now all FreeBSD needs is cloud-init so the resulting image can be
deployed and used in the cloud.