update brokeniso usage - revert F19 and F20 kickstart files
The iso command line issue seems to have been resolved between F17 and F18.

Further testing indicates that the improper use of the brokeniso flag
is the actual root cause of the issues with the original F19 JEOS ks
file.

With the flag set correctly we can now use the same kickstart file for
F18, F19 and F20.