if given an unknown new Fedora make best effort with latest known version info
This works roughly half of the time.  If, for example, we had added this code
for F18 it would have been possible for people to build F19 and F20 successfully
without fudging the TDL.  This is written in such a way that it does not confuse
ISO or modified media caching.