If given an unknown Fedora make best effort with latest.
This works roughly half of the time.  If, for example, we
had added this code for F18 it would have been possible for
people to build F19 and F20 successfully without fudging the
TDL.  This is written in such a way that it does not confuse
ISO or modified media caching.

Original patch for Ian Mcleod <imcleod@redhat.com>,
modified by me.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>