Use distro defaults for bootloader, partitioning and filesystems
Use the anaconda autopart to setup LVM partitions. It uses the distro defaults for
filesystems, eg RHEL 7 defaults to xfs, and also sets up the architecture specific
bootloader pieces correctly eg if the system is uEFI it adds the /boot/EFI
partition, on PPC it deals with PrEP and if specific x86 platforms need  bootbios
it deals with that as well. The type=lvm will give us the equivilent best LVM
layout that the old config provided as well so we're not second guessing the OS/HW.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>