Clean up config file searching.
As pointed out by jeffx, we didn't allow non-root
users to use the /etc/oz/oz.cfg file by default,
though there is no good reason not to.  Additionally
I noticed that we never allowed the root user to have
a private ~/.oz/oz.cfg file.  We fix these issues and
a couple of more issues with this commit (more info
about the algorithm is in the code comments).

Signed-off-by: Chris Lalancette <clalancette@gmail.com>