Remove a function that was basically duplicated.
It turns out that the _wait_for_guest_shutdown
and _wait_for_clean_shutdown methods were nearly
identical.  They differed only in how they dealt
with errors.  However, it turns out that the only
caller of wait_for_guest_shutdown didn't care
about the exception raised, and thus we can combine
the two and get rid of duplicated functionality.
The new code is a bit shorter, removes some
redundant code, and should be functionally equivalent
to the previous code.  The only exception to this
is that the code now waits a maximum of 10 seconds
instead of 20 for a clean shutdown.  Generally, I
think this is OK, but we can raise the limit if it
turns out that it is too short.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>