Deal with already mounted filesystems in guestfs.
btrfs volumes are a particular problem, in that different
snapshots could be mounted in the same place.  Thus,
we only mount one filesystem on one place in a generic
way.  Note that due to the length sorting we do, we'll
always mount the base filesystem, and not a snapshot,
which seems to work for us.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>